fix: edit unnecessary toast error msg on ceremony detail page(#230)
