+++ src/fsd_entities/ceremony/model/index.ts
@@ -1,3 +1,4 @@
-export { useCeremonyData } from './useAdminCeremonyData';
+export { useAdminCeremonyData } from './useAdminCeremonyData';
 export { useCeremonyCreateForm } from './useCeremonyCreateForm';
 export { useCeremonySettingForm } from './useCeremonySettingForm';
+export { useCeremonyData } from './useCeremonyDetailData';
