+++ src/fsd_entities/ceremony/model/useAdminCeremonyData.ts
@@ -8,19 +8,9 @@ import { ERROR_MESSAGES, MESSAGES } from '@/fsd_shared';
 
 import { getAdminCeremonyAwaitList, getAdminCeremonyDetail } from '../api';
 
-export const useCeremonyData = (ceremonyId?: string) => {
+export const useAdminCeremonyData = () => {
   const [ceremonyList, setCeremonyList] = useState<Ceremony.CeremonyItem[]>([]);
-  const [ceremonyDetails, setCeremonyDetails] = useState({
-    title: '',
-    type: '',
-    register: '',
-    content: '',
-    startDate: '',
-    endDate: '',
-    imageList: [] as string[],
-    applicantName: '',
-    applicantStudentId: '',
-  });
+
   const fetchCeremonyList = async () => {
     try {
       const data = await getAdminCeremonyAwaitList(0, 10);
@@ -34,32 +24,5 @@ export const useCeremonyData = (ceremonyId?: string) => {
     fetchCeremonyList();
   }, []);
 
-  useEffect(() => {
-    if (ceremonyId) {
-      const fetchCeremonyDetail = async () => {
-        try {
-          const cermonyContent = await getAdminCeremonyDetail(ceremonyId);
-          const matchedCeremony = ceremonyList.find((item) => item.id === ceremonyId);
-
-          setCeremonyDetails({
-            title: matchedCeremony ? matchedCeremony.title : cermonyContent.description,
-            type: cermonyContent.category,
-            register: cermonyContent.ceremonyState,
-            content: cermonyContent.description,
-            startDate: cermonyContent.startDate,
-            endDate: cermonyContent.endDate,
-            imageList: cermonyContent.attachedImageUrlList,
-            applicantName: cermonyContent.applicantName,
-            applicantStudentId: cermonyContent.applicantStudentId,
-          });
-        } catch (error) {
-          toast.error(`${MESSAGES.CEREMONY.DETAIL_CONTENT_TITLE} - ${ERROR_MESSAGES.DETAIL_CONTENT_FETCH_FAIL}`);
-        }
-      };
-
-      fetchCeremonyDetail();
-    }
-  }, [ceremonyId, ceremonyList]);
-
-  return { ceremonyList, ceremonyDetails };
+  return ceremonyList;
 };
